requirements.txt
django==5.0.6
gitpython
requests
dotenv
litellm[proxy]
openai-whisper